requirements.txt
numpy>=1.24.3
mss>=9.0.1
opencv-contrib-python==4.7.0.72
paddlepaddle==2.4.2
paddleocr==2.6.1.3
robotframework>=6.0.2
robotframework-dependencylibrary>=4.0.1
cairosvg>=2.7.0
Levenshtein==0.21.1
mouse >=0.7.1